Reversed VERTICAL index values for BoardPosition
Values were still incorrect according to official chess docs showing the board and how it is labelled. Position 8 is top right corner of board, which should be index 0 instead of index 7 and all subsequent values needed to be flipped to match.

Flipped the logic in pawn colors again back to what it was, as it seems the board and logic has corrected itself now.