CheckMate and InCheck completed!
Fixed an issue in Pawn Logic that allowed a pawn to move 2 squares with a piece in front of it, effectively jumping over the piece.

Completed logic for Is King in Check and for Check Mate as well!

Added Turn Description to make it easy to debug

Proudest commit of this project by far!