Model A approach
Model A suggests an interest approach that appears very simple. Have each ChessPiece track if it's an En Passant target. It tries to find the other ChessPiece using board.GetPieceAtPosition, which can be used to find another chess piece. If there is a piece there, check if it's an enpassant target. If it is, then a valid en passant move can be made. In the implement move, track the enpassant target for each ChessPiece. This can be done in the base class.

It assumes there exists a board.GetPieceAtPosition, which currently does not. Also, board.GetPieceAtPosition is very difficult to implement, because ChessBoard does not store a collection of ChessPiece; GameController is responsible for doing that. ChessBoard only represents the 2-dimensional integer array representing the Chess Board. It is a very primitive class that holds very few responsibilities. This makes it an unlikely candidate to ever implement such a method (at least without taking in a GameController as an additional argument). However, then that means all ChessPieces have to hold a reference to a GameController, and the GameController holds a collection of all ChessPieces. It's a circular reference that is a headache to deal with and makes this approach not very feasable now that I've gotten part way through it.