ETL For Automating Test Cases
Created new python project that acts like an ETL. It extracts replays from chess.com using their API to retrieve a list of PGN files. These PGN files are saved locally, then another process reads the files and converts the notation into chess notation my engine can understand.

There is still some kinks to be worked out (some pieces are showing as None in the output) but for the most part the piece name and position is captured, including the expected outcome. We can feed this into another test on the C# side that can read the file, queue up the inputs, and assert that the engine comes to the same conclusion about the game as it played out on chess.com

This uses python-chess library to manipulate the PGN data. So we have 3 different chess engine services interacting with each other: chess.com, python-chess, and this chess engine.

The python package dependencies are installed locally into this project. Though in the future might be better to use a requirements.txt file instead.