BoardID constructor, Bad Castle Exception fixed
Added a new constructor for ChessBoard that constructs a board with pieces based on a boardID string with unit test.

Added a call to generate boardID when a bad castle exception is thrown

Added strong typing to update_piece_tracking method in ETL. Still need to handle pawn promotions.

Made a public setter for TurnNumber in GameController. This was done so that a board state could be loaded up with a turn number preset. Useful for debugging broken board states.

In ChessPieceKing, added a check that the King has to be on its starting square in order for castling to be valid.

Added a TODO in KingCheckService, in that pieces should store what valid moves they can make, instead of this brute force approach. A pawn doesn't need to iterate over 64 squares to identify what valid moves it can make.

Added the boardID constructor unit test to ChessBoardTests.

Added a test for DebugBadCastleException - a certain kind of move would trigger a BadCastleException with this particular board state. This test helps address that issue.