Pawn Promotion in ETL fixed
This breaks many tests... And the implementation sucks. I'm not a fan of duplicating TurnNumber property in multiple classes.

Pawn Promotion has been addressed and fixed. But there is a long standing issue with En Passant... En Passant is only valid if it happens right after the pawn moves 2 squares... You cannot en passant many turns later. This engine doesn't check this constraint, creating odd scenarios where a pawn can capture 2 pieces in a single move (its regular capture + en passant). This scenario is not valid for chess, but the work required to fix this is not trivial at all.

Current implementation seems to fix this constraint, but now many tests are breaking.

Another day to work on this.