Big Logging Changes
Implemented more sophisticated logging for the codebase, with configuration that can turn on/off logging in specific areas of the codebase.

Enhanced ToStringTrait to allow for types to be skipped. Same with StaticLogger.

Created TestBase class for common setup/teardown code across all tests.

Fixed TestAnalyzer finally - it now finds the [TestNeeded] attributes across the codebase and generates a report.