Refactor Logging and En Passant Logic Patching
Refactored Logger so that when its dumping objects it only does so when the configuration flags are turned on.

Added a SimulationService which allows KingCheckService to simulate future turns without modifying the current game state. Needed for En Passant.

Added an enhanced MockConsoleService specifically for handling chess.com replays

Removed async/parallel processing of tests for GameController and GameManager, since these now rely on global data that could be modified unintentionally through race conditions, leading to bizarre test failures.

Enhanced Static Logger to support white list types - this way I can define which types I want to see logging for, allowing for more flexible control with logging.

Added LambdaQueue class, which is a service that invokes lambdas at a later time by queuing them. Also has a drain counter that keeps track of when a Lambda should be invoked. This is used for En Passant only.